+++ tiledb/sm/query/writer.h
@@ -340,6 +340,12 @@ class Writer {
   /** Sets config for query-level parameters only. */
   Status set_config(const Config& config);
 
+  /**
+   * Get the config of the writer
+   * @return Config
+   */
+  const Config& config() const;
+
   /** Sets current setting of check_coord_dups_ */
   void set_check_coord_dups(bool b);